Add option to override the default background executor, see #292
